Feature/item crud

+++ src/main/java/com/controle_de_gastos/notas_api/service/ItemService.java
@@ -1,17 +1,20 @@
 package com.controle_de_gastos.notas_api.service;
+import com.controle_de_gastos.notas_api.dto.projecao.AtributoSimplesProjecaoDTO;
+import com.controle_de_gastos.notas_api.dto.resposta.AtributoRespostaDTO;
+import com.controle_de_gastos.notas_api.dto.resposta.ItemComAtributoListRespostaDTO;
 import com.controle_de_gastos.notas_api.repository.ItemAtributoJuncaoRepository;
 import com.controle_de_gastos.notas_api.repository.ItemRepository;
 import com.controle_de_gastos.notas_api.repository.ItemTipoRepository;
 import com.controle_de_gastos.notas_api.repository.MarcaRepository;
 import com.controle_de_gastos.notas_api.dto.resposta.ItemRespostaDTO;
-import com.controle_de_gastos.notas_api.dto.resposta.ItemAtributosRespostaDTO;
 import com.controle_de_gastos.notas_api.dto.requisicao.ItemRequisicaoDTO;
 import com.controle_de_gastos.notas_api.model.Item;
-import com.controle_de_gastos.notas_api.model.ItemAtributoJuncao;
 import com.controle_de_gastos.notas_api.model.ItemTipo;
 import com.controle_de_gastos.notas_api.model.Marca;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import javax.swing.*;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,29 +38,22 @@ public class ItemService {
                 item.getPeso(),
                 item.getVersao(),
                 itemTipoService.toRespostaDTO(item.getTipo()),
-                marcaService.toRespostaDTO(item.getMarca())
+                marcaService.toRespostaDTO(item.getMarca()),
+                item.getItemAtributoJuncaos()
+                        .stream()
+                        .map(juncao -> new AtributoRespostaDTO(
+                                juncao.getAtributo().getId(),
+                                juncao.getAtributo().getNome()
+                        ))
+                        .toList()
         );
     }
 
-    public ItemAtributosRespostaDTO toItemComAtributoDTO(ItemAtributoJuncao itemAtributoJuncao){
-        return new ItemAtributosRespostaDTO(
-                itemAtributoJuncao.getItem().getId(),
-                itemAtributoJuncao.getAtributo().getId(),
-                itemAtributoJuncao.getAtributo().getNome()
-        );
-    }
 
     public List<ItemRespostaDTO> listartodos(){
         return itemRepository.findAll()
                 .stream()
                 .map(this::toRespostaDTO)
-                .toList();
-    }
-
-    public List<ItemAtributosRespostaDTO> listarAtributosPorItem(Integer id){
-        return itemAtributoJuncaoRepository.findByItemId(id)
-                .stream()
-                .map(this::toItemComAtributoDTO)
                 .toList();
     }
 
@@ -67,8 +63,48 @@ public class ItemService {
 
     }
 
+    public List<ItemComAtributoListRespostaDTO> listarAtributosPorItemTodos(){
+          return itemRepository.findAll()
+                  .stream()
+                  .map(item -> new ItemComAtributoListRespostaDTO(
+                          item.getId(),
+                          item.getNome(),
+                          item.getPeso(),
+                          item.getVersao(),
+                          item.getItemAtributoJuncaos()
+                                  .stream()
+                                  .map(juncao -> new AtributoSimplesProjecaoDTO(
+                                          juncao.getAtributo().getId(),
+                                          juncao.getAtributo().getNome()
+                                  )).toList()
 
-    public ItemRespostaDTO salvarItem(ItemRequisicaoDTO itemDTO){
+                  )).toList();
+    }
+
+    public Optional<ItemComAtributoListRespostaDTO> listarAtributosPorItemId(Integer id){
+        Optional<Item> item =  itemRepository.findById(id);
+        if(item.isEmpty()) return Optional.empty();
+
+        List<AtributoSimplesProjecaoDTO> atributos = itemAtributoJuncaoRepository.findByItemId(id)
+                .stream()
+                .map(juncao -> new AtributoSimplesProjecaoDTO(
+                        juncao.getAtributo().getId(),
+                        juncao.getAtributo().getNome()
+                ))
+                .toList();
+        return Optional.of(new ItemComAtributoListRespostaDTO(
+                item.get().getId(),
+                item.get().getNome(),
+                item.get().getPeso(),
+                item.get().getVersao(),
+                atributos
+        ));
+    }
+
+
+
+
+    public ItemRespostaDTO criarItem(ItemRequisicaoDTO itemDTO){
         ItemTipo tipo = itemTipoRepository.findById(itemDTO.idTipo())
                 .orElseThrow(()-> new RuntimeException("Tipo não encontrado"));
         Marca marca = marcaRepository.findById(itemDTO.idMarca())
@@ -85,7 +121,42 @@ public class ItemService {
         return toRespostaDTO(itemRepository.save(item));
     }
 
-    public void deletarPorID(Integer id){
-        itemRepository.deleteById(id);
+    public Optional<ItemRespostaDTO> atualizarTudo(Integer id, ItemRequisicaoDTO itemDTO){
+        Item item = itemRepository.findById(id).orElse(null);
+
+        if(item == null) return Optional.empty();
+
+        ItemTipo novoTipo = itemTipoRepository.findById(itemDTO.idTipo())
+                .orElseThrow(()->new IllegalStateException("Item não encontrado"));
+
+        Marca novaMarca = marcaRepository.findById(itemDTO.idMarca())
+                .orElseThrow(()-> new IllegalStateException("Marca não encontrada"));
+
+        if(!item.getTipo().getId().equals(novoTipo.getId())){
+            item.setTipo(novoTipo);
+        }
+
+        if (!item.getMarca().getId().equals(novaMarca.getId())){
+            item.setMarca(novaMarca);
+        }
+
+        item.setNome(itemDTO.nome());
+        item.setPeso(itemDTO.peso());
+        item.setVersao(itemDTO.versao());
+
+        return Optional.of(toRespostaDTO(itemRepository.save(item)));
+
+    }
+
+    public boolean deletarPorID(Integer id){
+        Optional<Item> item = itemRepository.findById(id);
+
+        if(item.isEmpty())  return false;
+
+        if(!item.get().getItemAtributoJuncaos().isEmpty() || !item.get().getNotaItemJuncaos().isEmpty()){
+            throw new RuntimeException("Não é possível deletar, existem dependências");
+        }
+                       itemRepository.deleteById(id);
+               return true;
     }
 }
